// types/restcountries.ts
export type RestCountries = RestCountry[];

export interface RestCountry {
  name: Name;
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: Currencies;
  idd: Idd;
  capital?: string[];
  altSpellings: string[];
  region: string;
  languages?: Languages;
  translations: Translations;
  latlng: number[];
  landlocked: boolean;
  area: number;
  demonyms?: Demonyms;
  flag: string;
  maps: Maps;
  population: number;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: string;
  capitalInfo: CapitalInfo;
  cioc?: string;
  subregion?: string;
  fifa?: string;
  borders?: string[];
  gini?: Gini;
  postalCode?: PostalCode;
}

export interface Name {
  common: string;
  official: string;
  nativeName?: NativeName;
}
// ...
export interface Currencies {
  XCD?: Xcd;
  NIO?: Nio;
  NZD?: Nzd;
  ILS?: Ils;
  EUR?: Eur;
  XAF?: Xaf;
  USD?: Usd;
  MNT?: Mnt;
  ZWL?: Zwl;
  GHS?: Ghs;
  AUD?: Aud;
  KID?: Kid;
  TOP?: Top;
  KMF?: Kmf;
  BND?: Bnd;
  SGD?: Sgd;
  KYD?: Kyd;
  PLN?: Pln;
  JMD?: Jmd;
  LRD?: Lrd;
  VES?: Ves;
  BTN?: Btn;
  INR?: Inr;
  MMK?: Mmk;
  BZD?: Bzd;
  RSD?: Rsd;
  BRL?: Brl;
  ALL?: All;
  IRR?: Irr;
  SSP?: Ssp;
  MVR?: Mvr;
  KWD?: Kwd;
  BSD?: Bsd;
  NGN?: Ngn;
  GMD?: Gmd;
  TJS?: Tjs;
  GTQ?: Gtq;
  SCR?: Scr;
  BMD?: Bmd;
  CLP?: Clp;
  BOB?: Bob;
  JPY?: Jpy;
  GBP?: Gbp;
  JEP?: Jep;
  WST?: Wst;
  AMD?: Amd;
  FKP?: Fkp;
  XOF?: Xof;
  MDL?: Mdl;
  KZT?: Kzt;
  AOA?: Aoa;
  MRU?: Mru;
  JOD?: Jod;
  KRW?: Krw;
  TZS?: Tzs;
  ERN?: Ern;
  NOK?: Nok;
  ANG?: Ang;
  PAB?: Pab;
  MZN?: Mzn;
  TMT?: Tmt;
  HKD?: Hkd;
  UYU?: Uyu;
  VND?: Vnd;
  GYD?: Gyd;
  TTD?: Ttd;
  RON?: Ron2;
  RWF?: Rwf;
  KES?: Kes;
  SHP?: Shp;
  VUV?: Vuv;
  XPF?: Xpf;
  CZK?: Czk;
  SOS?: Sos;
  CDF?: Cdf;
  SRD?: Srd;
  BHD?: Bhd;
  GNF?: Gnf;
  DKK?: Dkk;
  GIP?: Gip;
  PHP?: Php;
  BYN?: Byn;
  HNL?: Hnl;
  TVD?: Tvd;
  HTG?: Htg;
  THB?: Thb;
  CRC?: Crc;
  DZD?: Dzd;
  MAD?: Mad;
  ARS?: Ars;
  ISK?: Isk;
  LSL?: Lsl;
  ZAR?: Zar;
  MOP?: Mop;
  OMR?: Omr;
  ETB?: Etb;
  BAM?: Bam;
  SAR?: Sar;
  MWK?: Mwk;
  CAD?: Cad;
  STN?: Stn;
  MYR?: Myr;
  RUB?: Rub;
  KGS?: Kgs;
  CUC?: Cuc;
  CUP?: Cup;
  SEK?: Sek;
  IQD?: Iqd;
  KHR?: Khr;
  BGN?: Bgn;
  AED?: Aed;
  NAD?: Nad;
  IMP?: Imp;
  UAH?: Uah;
  PYG?: Pyg;
  SZL?: Szl;
  SYP?: Syp;
  MGA?: Mga;
  QAR?: Qar;
  DOP?: Dop;
  PGK?: Pgk;
  LBP?: Lbp;
  BWP?: Bwp;
  PKR?: Pkr;
  GEL?: Gel;
  NPR?: Npr;
  CHF?: Chf;
  FJD?: Fjd;
  SDG?: Sdg;
  SLL?: Sll;
  HUF?: Huf;
  TWD?: Twd;
  BBD?: Bbd;
  TND?: Tnd;
  IDR?: Idr;
  CVE?: Cve;
  LAK?: Lak;
  UGX?: Ugx;
  BIF?: Bif;
  LYD?: Lyd;
  MXN?: Mxn;
  MKD?: Mkd2;
  CNY?: Cny;
  YER?: Yer;
  GGP?: Ggp;
  SBD?: Sbd;
  FOK?: Fok;
  UZS?: Uzs;
  EGP?: Egp;
  LKR?: Lkr;
  BDT?: Bdt;
  PEN?: Pen;
  TRY?: Try;
  AFN?: Afn;
  AWG?: Awg;
  CKD?: Ckd;
  ZMW?: Zmw;
  AZN?: Azn;
  DJF?: Djf;
  KPW?: Kpw;
  MUR?: Mur;
  COP?: Cop;
}
// ...
export interface Idd {
  root?: string;
  suffixes?: string[];
}

export interface Languages {
  eng?: string;
  spa?: string;
  mri?: string;
  nzs?: string;
  ara?: string;
  heb?: string;
  deu?: string;
  fra?: string;
  nld?: string;
  kon?: string;
  lin?: string;
  smo?: string;
  por?: string;
  mon?: string;
  bwg?: string;
  kck?: string;
  khi?: string;
  ndc?: string;
  nde?: string;
  nya?: string;
  sna?: string;
  sot?: string;
  toi?: string;
  tsn?: string;
  tso?: string;
  ven?: string;
  xho?: string;
  zib?: string;
  gil?: string;
  ton?: string;
  cha?: string;
  zdj?: string;
  msa?: string;
  pol?: string;
  jam?: string;
  dzo?: string;
  mya?: string;
  bjz?: string;
  srp?: string;
  sqi?: string;
  fas?: string;
  div?: string;
  mlt?: string;
  rus?: string;
  tgk?: string;
  crs?: string;
  aym?: string;
  grn?: string;
  que?: string;
  jpn?: string;
  nrf?: string;
  hye?: string;
  ron?: string;
  kaz?: string;
  slk?: string;
  lit?: string;
  kor?: string;
  swa?: string;
  tir?: string;
  sag?: string;
  swe?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  pap?: string;
  gle?: string;
  tuk?: string;
  de?: string;
  zho?: string;
  ita?: string;
  lat?: string;
  vie?: string;
  tet?: string;
  est?: string;
  kin?: string;
  niu?: string;
  bis?: string;
  ces?: string;
  som?: string;
  lua?: string;
  dan?: string;
  fil?: string;
  pau?: string;
  lav?: string;
  bel?: string;
  tvl?: string;
  hat?: string;
  tha?: string;
  ber?: string;
  mey?: string;
  nau?: string;
  ltz?: string;
  isl?: string;
  mah?: string;
  slv?: string;
  cat?: string;
  eus?: string;
  glc?: string;
  amh?: string;
  bos?: string;
  hrv?: string;
  ell?: string;
  tur?: string;
  kir?: string;
  arc?: string;
  ckb?: string;
  khm?: string;
  kal?: string;
  bul?: string;
  afr?: string;
  her?: string;
  hgm?: string;
  kwn?: string;
  loz?: string;
  ndo?: string;
  glv?: string;
  ukr?: string;
  ssw?: string;
  cnr?: string;
  hin?: string;
  tam?: string;
  mlg?: string;
  nor?: string;
  pih?: string;
  hmo?: string;
  tpi?: string;
  urd?: string;
  kat?: string;
  nep?: string;
  fij?: string;
  hif?: string;
  gsw?: string;
  roh?: string;
  hun?: string;
  ind?: string;
  lao?: string;
  run?: string;
  nbl?: string;
  nso?: string;
  zul?: string;
  cal?: string;
  mkd?: string;
  nfr?: string;
  fao?: string;
  uzb?: string;
  sin?: string;
  ben?: string;
  prs?: string;
  pus?: string;
  rar?: string;
  fin?: string;
  tkl?: string;
  pov?: string;
  aze?: string;
  mfe?: string;
}

export interface Translations {
  ara: Ara2;
  bre: Bre;
  ces: Ces2;
  cym: Cym;
  deu: Deu2;
  est: Est2;
  fin: Fin2;
  fra: Fra2;
  hrv?: Hrv2;
  hun: Hun2;
  ita: Ita2;
  jpn?: Jpn2;
  kor: Kor2;
  nld: Nld2;
  per?: Per;
  pol: Pol2;
  por: Por2;
  rus: Rus2;
  slk: Slk2;
  spa: Spa2;
  srp: Srp2;
  swe: Swe2;
  tur: Tur2;
  urd: Urd2;
  zho?: Zho2;
}
// ...
export interface Demonyms {
  eng: Eng2;
  fra?: Fra3;
}

export interface Eng2 {
  f: string;
  m: string;
}

export interface Fra3 {
  f: string;
  m: string;
}

export interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

export interface Car {
  signs?: string[];
  side: string;
}

export interface Flags {
  png: string;
  svg: string;
  alt?: string;
}

export interface CoatOfArms {
  png?: string;
  svg?: string;
}

export interface CapitalInfo {
  latlng?: number[];
}

export interface Gini {
  "2014"?: number;
  "2016"?: number;
  "2018"?: number;
  "2011"?: number;
  "2019"?: number;
  "2006"?: number;
  "2015"?: number;
  "2004"?: number;
  "2017"?: number;
  "1999"?: number;
  "2013"?: number;
  "2009"?: number;
  "2010"?: number;
  "2008"?: number;
  "1998"?: number;
  "1992"?: number;
  "2012"?: number;
  "2003"?: number;
  "2005"?: number;
}

export interface PostalCode {
  format: string;
  regex?: string;
}